extras/locust/requirements.txt
###### Requirements without Version Specifiers ######`
locust
locust-plugins
grpcio
grpcio-tools
asyncio